Git ignore added, starting work on main Webapp
For the time being the _OLD assets are still part of the project, in
order to keep the original scripts of the project.

Started main work on CSS folder and overall styles, HTML setup is also
there.

Instructions for network setup can be found in the receiver.js folder

Have fun!